chore: merge staging (#465)
* feat(discord-bot): integrate with spartan-ecdsa (#462)

* chore: remove uneeded deps/files

* chore: fix deps

* feat: use en vars for VERIFICATION_CHANNEL_ID and VERIFIED_ROLE_ID

* refactor: change formatting of messages sent by bot

link to channels
use different font style

* chore: remove unused deps

* feat: verify spartan proof with discord bot

* fix: bump @types/node

Need > 20 to have fetch typings

* chore: update comment

* ci: remove debug step

* docs: update .env.example

* docs: update discord/bot/README

* refactor: update error message if missing env var

* fix(ui): do not pre render /api/verify route

* refactor: use verifier object to load wasm async

* fix: link wasm files to `server/static`

webpack 5 breaks dynamic import of wasm
https://github.com/vercel/next.js/issues/25852

* ci: link wasm files to `.next/server/static` in Dockerfile

* chore: debug

* fix(ui): set CORS headers

* refactor: add info logs

* ci: refactor Dockerfile and fly toml

* fix: remove remaining merge hints

* feat: restrict verification request to discord bot origin (#466)

* refactor: add info logs

* wip

* fix  import

* refactor: remove wallet connect id from fly config

Set already in Dockerfile

* ci: refactor setting of env vars in Dockerfile and fly configs (#469)

ci: update dockerfiles and fly env vars

+++ discord-bot/src/events/MessageCreate.ts
@@ -1,8 +1,11 @@
 import { config } from '#/config'
-import { verifyOnChain } from '#/verify-onchain'
-import { Attachment, Collection, Events, Message, TextChannel } from 'discord.js'
+import { info } from '#/logger'
+import { Attachment, Collection, Events, Message, PermissionFlagsBits, TextChannel } from 'discord.js'
 import { _Event } from './_Event'
 import { HandledEvent } from './interface'
+
+const VERIFICATION_URL = `${config.urls.ui}/api/verify`
+const DELETE_DELAY = 10_000
 
 export class MessageCreate extends _Event {
   override bind = true
@@ -16,27 +19,34 @@ export class MessageCreate extends _Event {
     if (channelName !== `private-verify-${message.author.username}`) return
     if (message.attachments.size === 0) return
 
-    const { proof, publicSignals } = await this._handleAttachments(
+    info(`Fetching proof attachment uploaded by ${message.author.username}`)
+    const proof = await this._handleProofBinaryAttachment(
       message.attachments,
     )
 
-    // FIXME: isn't viem supposed to give the typings automagically from the JSON ABIs?
-    const valid: boolean = await verifyOnChain({
-      // FIXME
-      // @ts-expect-error
-      proof,
-      // FIXME
-      // @ts-expect-error
-      publicSignals,
+    info(`Posting ${message.author.username}'s binary proof at ${VERIFICATION_URL} for verification`)
+    const res = await fetch(VERIFICATION_URL, {
+      body: proof,
+      headers: {
+        [config.auth.header]: config.auth.secret,
+        'Content-Type': 'application/octet-stream',
+      },
+      method: 'POST',
     })
+    const { status, statusText } = res
+    if (status !== 200) throw new Error(`${status} ${statusText}`)
 
-    if (valid) {
+    const valid = await res.json()
+    if (valid === true) {
       if (message.member === null || message.member === undefined)
         throw new Error('No member found')
+
       await message.member.roles.add(config.VERIFIED_ROLE_ID)
+      info(`Granted ${message.author.username} the verified role`)
+
       await message.channel.send({
         content:
-          `Congrats \`${message.author.username}\`, you proof is valid ✅! You have been granted the verified role. This private channel will be deleted in 10s.`,
+          `Congrats **${message.author.username}**, you proof is valid ✅! You have been granted the *verified* role. This private channel will be deleted in 10s.`,
       })
 
       setTimeout(() => {
@@ -45,52 +55,45 @@ export class MessageCreate extends _Event {
 
           const verificationChannel = message.guild?.channels.cache.get(
             config.VERIFICATION_CHANNEL_ID,
-          ) as TextChannel
+          )
+          // TODO perform these checks once when initializing the bot?
+          if (!(verificationChannel instanceof TextChannel)) throw new Error('No verification text channel found')
+          const permissions = verificationChannel.permissionsFor(message.client.user)
+          if (permissions === null) {
+            throw new Error(
+              `No permissions found for bot ${message.client.user.username} in #${verificationChannel.name}`,
+            )
+          }
+          if (!permissions.has(PermissionFlagsBits.ReadMessageHistory)) {
+            throw new Error(
+              `Bot ${message.client.user.username} does not have permission to read message history in #${verificationChannel.name}`,
+            )
+          }
+
           const pastMessages = await verificationChannel.messages.fetch()
           const botMessage = pastMessages.find(
             (pastMessage: Message) =>
               pastMessage.author.id === config.CLIENT_ID
               && pastMessage.content.includes(
-                `Hello \`${message.author.username}\`, please check #private-verify-${message.author.username} for further instructions.`,
+                message.author.username,
               ),
           )
           if (botMessage !== undefined) await botMessage.delete()
-        })().catch((err) => {
-          console.error(err)
-        })
-      }, 10000)
+        })()
+      }, DELETE_DELAY)
     } else {
       await message.channel.send({
         content:
-          `Sorry \`${message.author.username}\`, your proof is invalid ❌. You have not been granted the verified role.`,
+          `Sorry **${message.author.username}**, your proof is invalid ❌. You have not been granted the *verified* role.`,
       })
     }
   }
 
-  private async _handleAttachments(
-    attachments: Collection<string, Attachment>,
-  ) {
-    const [proof, publicSignals] = await Promise.all(
-      ['proof.json', 'public.json'].map(async (name) => this._handleJsonAttachment({ attachments, name })),
-    )
-
-    return {
-      proof,
-      publicSignals,
-    }
-  }
-
-  private async _handleJsonAttachment({
-    attachments,
-    name,
-  }: {
-    attachments: Collection<string, Attachment>
-    name: string
-  }) {
-    const attachment = attachments.find((a) => a.name === name)
+  private async _handleProofBinaryAttachment(attachments: Collection<string, Attachment>) {
+    const attachment = attachments.find((a) => a.name === config.proofAttachmentName)
     if (attachment === undefined)
-      throw new Error(`No ${name} attachment in message`)
+      throw new Error(`No \`${config.proofAttachmentName}\` attachment in message`)
 
-    return fetch(attachment.url).then(async (res) => res.json())
+    return fetch(attachment.url).then(res => res.arrayBuffer())
   }
 }
